refactor examples #219
