+++ _examples/1_get_recent_activity/activity.go
