+++ _examples/1_get_recent_activity/go.mod
@@ -1,0 +1,5 @@
+module gotwi-example-1
+
+go 1.18
+
+require github.com/michimani/gotwi v0.10.1
