+++ _examples/1_get_recent_activity/main.go
@@ -1,0 +1,40 @@
+package main
+
+import (
+	"fmt"
+	"os"
+
+	"github.com/michimani/gotwi"
+)
+
+const (
+	OAuthTokenEnvKeyName       = "GOTWI_ACCESS_TOKEN"
+	OAuthTokenSecretEnvKeyName = "GOTWI_ACCESS_TOKEN_SECRET"
+)
+
+func main() {
+	args := os.Args
+	if len(args) < 2 {
+		fmt.Println("The 1st parameter is required for your account id.")
+		os.Exit(1)
+	}
+
+	accountID := args[1]
+
+	oauth1Client, err := newOAuth1Client()
+	if err != nil {
+		panic(err)
+	}
+
+	onlyFollowsRecentActivity(oauth1Client, accountID)
+}
+
+func newOAuth1Client() (*gotwi.GotwiClient, error) {
+	in := &gotwi.NewGotwiClientInput{
+		AuthenticationMethod: gotwi.AuthenMethodOAuth1UserContext,
+		OAuthToken:           os.Getenv(OAuthTokenEnvKeyName),
+		OAuthTokenSecret:     os.Getenv(OAuthTokenSecretEnvKeyName),
+	}
+
+	return gotwi.NewGotwiClient(in)
+}
