+++ _examples/2_post_delete_tweet/go.mod
@@ -1,0 +1,5 @@
+module gotwi-example-2
+
+go 1.18
+
+require github.com/michimani/gotwi v0.10.1
