+++ _examples/2_post_delete_tweet/tweets.go
