--- _examples/go.mod
@@ -1,7 +1,0 @@
-module github.com/michimani/gotwi/examples
-
-go 1.18
-
-require github.com/michimani/gotwi v0.10.0
-
-// replace github.com/michimani/gotwi v0.10.0 => ../
